locations の useState をカスタムフックに切り出し
