+++ src/utils/types.ts
@@ -1,0 +1,10 @@
+export type locationType = {
+  id: string;
+  location: {
+    latitude: number;
+    longitude: number;
+    altitude: number;
+    name: string;
+  };
+  enabledNotify: boolean;
+};
